ONYX-43855: Refactor edge controllers to move logic to domain
